projects/llm-deepteam-eval/pyproject.toml
[project]
name = "llm-deepteam-eval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepeval>=3.4.9",
    "deepteam>=0.2.5",
    "google-generativeai>=0.8.5"
]
